+++ src/sql/command.py
@@ -11,7 +11,9 @@ from sql.util import validate_nonidentifier_connection
 
 class SQLPlotCommand:
     def __init__(self, magic, line) -> None:
-        self.args = parse.magic_args(magic.execute, line)
+        self.args = parse.magic_args(
+            magic.execute, line, "sqlplot", allowed_duplicates=["-w", "--with"]
+        )
 
 
 class SQLCommand:
@@ -24,7 +26,13 @@ class SQLCommand:
         self._line = line
         self._cell = cell
 
-        self.args = parse.magic_args(magic.execute, line)
+        self.args = parse.magic_args(
+            magic.execute,
+            line,
+            "sql",
+            allowed_duplicates=["-w", "--with", "--append", "--interact"],
+        )
+
         # self.args.line (everything that appears after %sql/%%sql in the first line)
         # is split in tokens (delimited by spaces), this checks if we have one arg
         one_arg = len(self.args.line) == 1