+++ src/sql/magic_cmd.py
@@ -13,6 +13,7 @@ from sql.cmd.explore import explore
 from sql.cmd.snippets import snippets
 from sql.cmd.connect import connect
 from sql.connection import ConnectionManager
+from sql.util import check_duplicate_arguments
 
 try:
     from traitlets.config.configurable import Configurable
@@ -74,6 +75,17 @@ class SqlCmdMagic(Magics, Configurable):
             # directly use shlex since SqlCmdMagic does not use magic_args from parse.py
             split = shlex.split(line, posix=False)
             command, others = split[0].strip(), split[1:]
+            if others:
+                check_duplicate_arguments(
+                    self.execute,
+                    "sqlcmd",
+                    split,
+                    disallowed_aliases={
+                        "-t": "--table",
+                        "-s": "--schema",
+                        "-o": "--output",
+                    },
+                )
 
             if command in AVAILABLE_SQLCMD_COMMANDS:
                 if (