+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * [Fix] Fix bug causing empty result on SQL with trailing semicolon and comment (#907)
 * [Fix] Fix bug %sql not parsing JSON arrow operators correctly (#918)
 * [Fix] Fixed bug that returns empty results when exception is raised from DB driver
+* [Fix] Added guards to check and raise errors when arguments are entered twice in %sql, %sqlcmd and %sqlplot (#806)
 * [Fix] Fixed bug that returns snippet typo error message when another table is misspelled (#940)
 
 ## 0.10.3 (2023-11-06)